[QA] 폰트 교체 (#160)
fix : 폰트 교체 및 프리뷰 기능 androidMain으로 변경
