+++ core/designsystem/src/commonMain/kotlin/com/whatever/caramel/core/designsystem/themes/CaramelTypography.kt
@@ -245,46 +245,3 @@ data class CaramelTypography(
         )
     }
 }
-
-@Composable
-fun CaramelTypographyPreview() {
-    CaramelTheme {
-        Column(
-            modifier = Modifier.fillMaxSize(),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.spacedBy(3.dp, Alignment.CenterVertically)
-        ) {
-            val styles = mapOf(
-                CaramelTheme.typography.display to "Display",
-                CaramelTheme.typography.heading1 to "heading1",
-                CaramelTheme.typography.heading2 to "heading2",
-                CaramelTheme.typography.heading3 to "heading3",
-                CaramelTheme.typography.body1.bold to "body1.bold",
-                CaramelTheme.typography.body1.regular to "body1.regular",
-                CaramelTheme.typography.body2.bold to "body2.bold",
-                CaramelTheme.typography.body2.regular to "body2.regular",
-                CaramelTheme.typography.body2.reading to "body2.reading",
-                CaramelTheme.typography.body3.bold to "body3.bold",
-                CaramelTheme.typography.body3.regular to "body3.regular",
-                CaramelTheme.typography.body3.reading to "body3.reading",
-                CaramelTheme.typography.body4.bold to "body4.bold",
-                CaramelTheme.typography.body4.regular to "body4.regular",
-                CaramelTheme.typography.label1.bold to "label1.bold",
-                CaramelTheme.typography.label1.regular to "label1.regular",
-                CaramelTheme.typography.label2.bold to "label2.bold",
-                CaramelTheme.typography.label2.regular to "label2.regular",
-                CaramelTheme.typography.label3.bold to "label3.bold",
-                CaramelTheme.typography.label3.regular to "label3.regular",
-            )
-
-            for (style in styles.entries) {
-                Text(
-                    text = "Whatever x Caramel \n" + "${style.value}",
-                    textAlign = TextAlign.Center,
-                    style = style.key,
-                    color = Color.Black
-                )
-            }
-        }
-    }
-}
